requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
boto3==1.24.32
botocore==1.27.32
charset-normalizer==2.1.0
frozenlist==1.3.0
idna==3.3
jmespath==1.0.1
multidict==6.0.2
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
smart-open==6.0.0
urllib3==1.26.10
yarl==1.7.2
